update: Refactored Flyer Results Screen
the types.ts was created for the specific values in this page and the
index.ts is for easier imports
